Optionally save API token in system keyring
Closes #2. This commit removes the options to save the API token in an
environment variable or in a plaintext file because those options were
less secure than this new one, and I'm not sure they were necessary.
Maybe I will add one or both of them back in the future.

Zalando/go-keyring was chosen over 99designs/keyring because
the latter has not had any updates in 3 years, and for the same reason,
GitHub's CLI app (`gh`) devs chose the former and specifically not the
latter. Version 0.2.5 of zalando/go-keyring was chosen because of what
was discussed in https://github.com/cli/cli issue 9819.